+++ lib/ex/parser/subscanners.py
@@ -16,6 +16,7 @@ from .scanner_command_edit import scan_command_edit
 from .scanner_command_exit import scan_command_exit
 from .scanner_command_file import scan_command_file
 from .scanner_command_global import scan_command_global
+from .scanner_command_help import scan_command_help
 from .scanner_command_let import scan_command_let
 from .scanner_command_map import scan_command_map
 from .scanner_command_move import scan_command_move
@@ -70,6 +71,7 @@ patterns[r'd(?:elete)?'] = scan_command_delete
 patterns[r'exi(?:t)?'] = scan_command_exit
 patterns[r'f(?:ile)?'] = scan_command_file
 patterns[r'g(?:lobal)?(?=[^ ])'] = scan_command_global
+patterns[r'h(?:elp)?'] = scan_command_help
 patterns[r'(?:ls|files|buffers)!?'] = scan_command_buffers
 patterns[r'vs(?:plit)?'] = scan_command_vsplit
 patterns[r'x(?:it)?$'] = scan_command_exit